+++ src/main/java/co/smok/cassandra/tracing/EmptyTraceState.java
@@ -2,7 +2,6 @@ package co.smok.cassandra.tracing;
 
 import org.apache.cassandra.locator.InetAddressAndPort;
 import org.apache.cassandra.tracing.Tracing;
-import org.apache.cassandra.utils.TimeUUID;
 
 public class EmptyTraceState extends CommonTraceState {