+++ src/main/java/co/smok/cassandra/tracing/JaegerTracing.java
@@ -1,17 +1,9 @@
 package co.smok.cassandra.tracing;
 
-import ch.qos.logback.core.pattern.color.BoldYellowCompositeConverter;
 import io.jaegertracing.internal.JaegerSpan;
 import io.jaegertracing.internal.JaegerSpanContext;
-import io.jaegertracing.internal.JaegerTracer;
-import io.jaegertracing.internal.propagation.BinaryCodec;
-import io.opentracing.References;
-import io.opentracing.Span;
 import io.opentracing.propagation.Binary;
-import io.opentracing.propagation.BinaryExtract;
-import io.opentracing.propagation.BinaryInject;
 import io.opentracing.propagation.Format;
-import io.opentracing.tag.Tags;
 import org.apache.cassandra.concurrent.ExecutorLocals;
 import org.apache.cassandra.locator.InetAddressAndPort;
 import org.apache.cassandra.net.Message;
@@ -19,28 +11,21 @@ import org.apache.cassandra.net.ParamType;
 import org.apache.cassandra.tracing.TraceState;
 import org.apache.cassandra.tracing.Tracing;
 import org.apache.cassandra.utils.TimeUUID;
-import org.jetbrains.annotations.NotNull;
-import org.psjava.ds.array.Array;
-import org.psjava.ds.array.DynamicArray;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
 import java.net.InetAddress;
 import java.nio.ByteBuffer;
-import java.nio.charset.StandardCharsets;
-import java.sql.Time;
-import java.time.Instant;
-import java.util.*;
-import java.util.concurrent.ConcurrentHashMap;
-import java.util.concurrent.ConcurrentMap;
+import java.util.HashMap;
+import java.util.Map;
 
 /**
  * These two come in two flavors: coordinator-based traces, created by newSession,
  * and child traces, created by initializeFromMessage.
- *
+ * <p>
  * This has two methods - get() and set(), which are thread-local, and a sessions, which is a map of UUID to traces,
  * and is NOT thread-local.
- *
+ * <p>
  * There is only one instance of this class, and it serves as a factory for TraceStates
  */
 public final class JaegerTracing extends Tracing {
@@ -52,60 +37,23 @@ public final class JaegerTracing extends Tracing {
      * This class has two purposes:
      * 1. a replacement for Tracing.sessions more adequate to our needs
      * 2. a routing table for responses to route them to correct JaegerTraceStates
-      */
+     */
 
-    private static class SpanContextMap {
-        final private Map<String, JaegerTraceState> map = new HashMap<>();
-
-        private String contextToString(JaegerSpanContext context) {
-            return context.getTraceId()+":"+context.getSpanId();
-        }
-
-        public void put(JaegerSpanContext context, JaegerTraceState jts) {
-            this.map.put(contextToString(context), jts);
-        }
-
-        public void put(JaegerTraceState state) {
-            this.map.put(contextToString(state.parentSpan.context()), state);
-        }
-
-        public void put(JaegerSpan span, JaegerTraceState jts) {
-            this.map.put(contextToString(span.context()), jts);
-        }
-
-        public JaegerTraceState get(JaegerSpanContext context) {
-            return this.map.get(contextToString(context));
-        }
-
-        public JaegerTraceState pop(JaegerSpanContext context) {
-            String ctxt = contextToString(context);
-            JaegerTraceState state = this.map.get(ctxt);
-            this.map.remove(ctxt);
-            return state;
-        }
-
-        public void remove(JaegerSpanContext context) {
-            this.map.remove(contextToString(context));
-        }
-        public void remove(JaegerSpan span) {
-            this.map.remove(contextToString(span.context()));
-        }
-    }
 
     // a replacement for Tracing.sessions - used to store all sessions, coordinator or replicas, but only those that we're responsible for
-    final private JaegerTracing.SpanContextMap my_map = new JaegerTracing.SpanContextMap();
-
-    final private JaegerTracing.SpanContextMap routing_table = new JaegerTracing.SpanContextMap();
+    final private SpanContextMap my_map = new SpanContextMap();
+    // sub-spans of coordinator's parent span may issue requests of their own, we need to track them
+    final private SpanContextMap routing_table = new SpanContextMap();
 
     private static final InetAddressAndPort localHost = InetAddressAndPort.getLocalHost();
+
     /* a empty constructor is necessary for Cassandra to initialize this class **/
     public JaegerTracing() {
     }
 
     @Override
-    public CommonTraceState get()
-    {
-        return (CommonTraceState)ExecutorLocals.current().traceState;
+    public CommonTraceState get() {
+        return (CommonTraceState) ExecutorLocals.current().traceState;
     }
 
     /**
@@ -116,36 +64,22 @@ public final class JaegerTracing extends Tracing {
     protected void stopSessionImpl() {
     }
 
-    /**
-     * Register that a particular context should be mapped to this coordinator state
-     */
-    protected void registerRoutingTable(JaegerTraceState state, JaegerSpanContext context) {
-        this.routing_table.put(context, state);
-    }
-
     private boolean nullOrEmpty(CommonTraceState state) {
         return (state == null) || (state.isEmpty());
-    }
-
-    public CommonTraceState get(JaegerSpanContext context) {
-        return this.my_map.get(context);
     }
 
     @Override
     /**
      * Called by coordinator when it's time to stop shop.
      */
-    public void stopSession()
-    {
-        CommonTraceState state = get();
-        if (this.nullOrEmpty(state))
-        {
+    public void stopSession() {
+        CommonTraceState state = this.get();
+        if (this.nullOrEmpty(state)) {
             return;
         }
-        JaegerTraceState jts = (JaegerTraceState)state;
+        JaegerTraceState jts = (JaegerTraceState) state;
         jts.stop();
         my_map.remove(jts.parentSpan);
-        logger.warn("Stopping coordinator session "+jts.parentSpan.context());
         set(null);
     }
 
@@ -158,11 +92,11 @@ public final class JaegerTracing extends Tracing {
      * @param parameters likely to contain info about queries
      */
     public TraceState begin(String request, InetAddress client, Map<String, String> parameters) {
-        CommonTraceState ts = get();
+        CommonTraceState ts = this.get();
         if (this.nullOrEmpty(ts)) {
             return ts;
         }
-        JaegerTraceState state = (JaegerTraceState)ts;
+        JaegerTraceState state = (JaegerTraceState) ts;
 
         state.parentSpan.setTag("request", request);
         String query = parameters.get("query");
@@ -172,7 +106,6 @@ public final class JaegerTracing extends Tracing {
         if (client != null) {
             state.parentSpan.setTag("client", client.toString());
         }
-        logger.warn("Started a coordinator session with "+state.parentSpan.context().toString());
         return state;
     }
 
@@ -183,24 +116,20 @@ public final class JaegerTracing extends Tracing {
     }
 
     @Override
-     // This is called by the thread that holds the current state, so we can safely attach it.
+    // This is called by the thread that holds the current state, so we can safely attach it.
     public Map<ParamType, Object> addTraceHeaders(Map<ParamType, Object> addToMutable) {
         CommonTraceState ts = get();
         if (this.nullOrEmpty(ts)) {
             return addToMutable;
         }
-        JaegerTraceState state = (JaegerTraceState)ts;
-        JaegerTracing.BinaryExtractor be = new JaegerTracing.BinaryExtractor();
+        JaegerTraceState state = (JaegerTraceState) ts;
+        BinaryExtractor be = new BinaryExtractor();
         JaegerSpanContext context_to_attach = state.getContextToAttach();
         JaegerTracingSetup.tracer.inject(context_to_attach, Format.Builtin.BINARY, be);
 
         if (state.isCoordinator) {
-            logger.warn("I am a coordinator, my parentContext is "+state.parentSpan.context().toString()+" Im attaching "+context_to_attach.toString());
             this.routing_table.put(state.getContextToAttach(), state);
-        } else {
-            logger.warn("I am a replica, Im attaching "+context_to_attach.toString());
         }
-
         addToMutable.put(ParamType.CUSTOM_MAP, be.toCustomMap());
         return addToMutable;
     }
@@ -211,7 +140,6 @@ public final class JaegerTracing extends Tracing {
      * Called for non-local traces (traces that are not initiated by local node == coordinator).
      */
     public void trace(final ByteBuffer sessionId, final String message, final int ttl) {
-        logger.warn("trace() non-local got called with"+message);
     }
 
     @Override
@@ -219,8 +147,7 @@ public final class JaegerTracing extends Tracing {
      * This is called by the replica when sending a message to coordinator, or by coordinator when sending to a replica.
      * But after headers have been attached
      */
-    public void traceOutgoingMessage(Message<?> message, int serializedSize, InetAddressAndPort sendTo)
-    {
+    public void traceOutgoingMessage(Message<?> message, int serializedSize, InetAddressAndPort sendTo) {
         // We can't do much here, because we don't know who sent this message
     }
 
@@ -232,7 +159,7 @@ public final class JaegerTracing extends Tracing {
      * This is allowed to return a null, but not allowed to set a thread-local state
      */
     public TraceState initializeFromMessage(final Message.Header header) {
-        final BinaryExtractor be = new JaegerTracing.BinaryExtractor(header.customParams());
+        final BinaryExtractor be = new BinaryExtractor(header.customParams());
         if (be.isEmpty()) {
             return null;
         }
@@ -245,10 +172,9 @@ public final class JaegerTracing extends Tracing {
             // It's a response, let's go find the responsible JaegerTraceState for it
             JaegerTraceState parent = this.routing_table.pop(context);
             if (parent == null) {
-                logger.warn("Received an invalid response type "+header.verb.toString()+" for context "+context.toString());
+                logger.info("Received a timeouted replica for context " + context.toString());
                 return null;
             }
-            logger.warn("Received a response for parent "+parent.parentSpan.context().toString()+" replicas contexts was "+context.toString());
             parent.responseReceived();
             return parent;
         } else {
@@ -257,7 +183,7 @@ public final class JaegerTracing extends Tracing {
             JaegerTraceState trace = JaegerTraceState.asFollower(header.from, header.traceType(), context);
             my_map.put(trace);
             return trace;
-       }
+        }
     }
 
     @Override
@@ -269,8 +195,7 @@ public final class JaegerTracing extends Tracing {
     /**
      * This is called only on coordinator node
      */
-    protected TimeUUID newSession(TimeUUID sessionId, TraceType traceType, Map<String,ByteBuffer> customPayload)
-    {
+    protected TimeUUID newSession(TimeUUID sessionId, TraceType traceType, Map<String, ByteBuffer> customPayload) {
         assert get() == null;
 
         final StandardTextMap stm = new StandardTextMap(customPayload);
@@ -293,56 +218,8 @@ public final class JaegerTracing extends Tracing {
     }
 
     @Override
-    /**
-     * Called on followers, when they're through with their jobs.
-     */
-    public void doneWithNonLocalSession(TraceState state)
-    {
-        CommonTraceState cts = (CommonTraceState)state;
-        if (this.nullOrEmpty(cts)) {
-            return;
-        }
-        JaegerTraceState jts = (JaegerTraceState)cts;
-        jts.stop();
-        logger.warn("We are through with "+jts.parentSpan.context().toString());
-        my_map.remove(jts.parentSpan);
+    public void doneWithNonLocalSession(TraceState state) {
     }
 
-    static private class BinaryExtractor implements Binary {
 
-        private ByteBuffer buffer = null;
-
-        public boolean isEmpty() {
-            return this.buffer == null;
-        }
-
-        public BinaryExtractor() {}
-
-        public BinaryExtractor(Map<String, byte[]> map) {
-            if (map == null) {
-                return;
-            }
-            if (!map.containsKey(JaegerTracingSetup.trace_key)) {
-                return;
-            }
-            this.buffer = ByteBuffer.wrap(map.get(JaegerTracingSetup.trace_key)).asReadOnlyBuffer();
-        }
-
-        @Override
-        public ByteBuffer extractionBuffer() {
-            return this.buffer;
-        }
-
-        @Override
-        public ByteBuffer injectionBuffer(int length) {
-            this.buffer = ByteBuffer.allocate(length);
-            return this.buffer;
-        }
-
-        public Map<String, byte[]> toCustomMap() {
-            Map<String, byte[]> map = new HashMap<>();
-            map.put(JaegerTracingSetup.trace_key, this.buffer.array());
-            return map;
-        }
-    }
 }